Improved index page
Centered the content.
Added file selection handling.
Added a span to show the name of the file selected. Made it responsive.
Made Import button show the contents of the file.
Changed button styles to reflect whether file is selected or not. By
default Import button is hidden since there is no file selected.
Added a spinner to indicate that the "Import is in progress". Added
"font awesome" library.